+++ spessasynth_core/src/synthesizer/audio_engine/effects/insertion/ph_auto_wah.c
@@ -15,9 +15,9 @@ typedef struct {
 	SS_InsertionProcessor base;
 	SS_PhaserFX phaser;
 	SS_AutoWahFX auto_wah;
-	float ph_pan; /* 0..127, index into pan_table */
-	float aw_pan;
-	float level;
+	double ph_pan; /* 0..127, index into pan_table */
+	double aw_pan;
+	double level;
 	float *buf_ph;
 	float *buf_aw;
 	uint32_t buf_size;
@@ -29,9 +29,9 @@ static void phaw_process(SS_InsertionProcessor *self,
                          float *oRev, float *oCho, float *oDel,
                          int start, int n) {
 	SS_PhAutoWahFX *e = (SS_PhAutoWahFX *)self;
-	float rev = self->send_level_to_reverb;
-	float cho = self->send_level_to_chorus;
-	float del = self->send_level_to_delay;
+	double rev = self->send_level_to_reverb;
+	double cho = self->send_level_to_chorus;
+	double del = self->send_level_to_delay;
 
 	/* Process phaser (only left input) into buf_ph */
 	memset(e->buf_ph, 0, sizeof(float) * n);
@@ -53,18 +53,18 @@ static void phaw_process(SS_InsertionProcessor *self,
 	int aw_idx = (int)e->aw_pan;
 	if(aw_idx < 0) aw_idx = 0;
 	if(aw_idx > 127) aw_idx = 127;
-	float phL = ss_pan_table_left[ph_idx], phR = ss_pan_table_right[ph_idx];
-	float awL = ss_pan_table_left[aw_idx], awR = ss_pan_table_right[aw_idx];
+	double phL = ss_pan_table_left[ph_idx], phR = ss_pan_table_right[ph_idx];
+	double awL = ss_pan_table_left[aw_idx], awR = ss_pan_table_right[aw_idx];
 
 	for(int i = 0; i < n; i++) {
 		/* Divide by 2: each processor mixed both L+R into one buffer */
-		float out_ph = e->buf_ph[i] * 0.5f * e->level;
-		float out_aw = e->buf_aw[i] * 0.5f * e->level;
-		float outL = out_ph * phL + out_aw * awL;
-		float outR = out_ph * phR + out_aw * awR;
+		double out_ph = e->buf_ph[i] * 0.5 * e->level;
+		double out_aw = e->buf_aw[i] * 0.5 * e->level;
+		double outL = out_ph * phL + out_aw * awL;
+		double outR = out_ph * phR + out_aw * awR;
 		oL[start + i] += outL;
 		oR[start + i] += outR;
-		float mono = (outL + outR) * 0.5f;
+		double mono = (outL + outR) * 0.5;
 		if(oRev) oRev[i] += mono * rev;
 		if(oCho) oCho[i] += mono * cho;
 		if(oDel) oDel[i] += mono * del;
@@ -95,7 +95,7 @@ static void phaw_set_param(SS_InsertionProcessor *self, int p, int v) {
 			ss_auto_wah_set_param(&e->auto_wah.base, 0x16, v);
 			break;
 		case 0x16:
-			e->level = (float)v / 127.0f;
+			e->level = (float)v / 127.0;
 			break;
 		default:
 			break;
@@ -106,7 +106,7 @@ static void phaw_reset(SS_InsertionProcessor *self) {
 	SS_PhAutoWahFX *e = (SS_PhAutoWahFX *)self;
 	e->ph_pan = 0;
 	e->aw_pan = 127;
-	e->level = 1.0f;
+	e->level = 1.0;
 	ss_phaser_reset(&e->phaser.base);
 	ss_auto_wah_reset(&e->auto_wah.base);
 	/* Override sub-processor levels to full */
@@ -137,7 +137,7 @@ SS_InsertionProcessor *ss_insertion_ph_auto_wah_create(uint32_t type, uint32_t s
 	e->buf_size = max_buf_size;
 
 	e->base.type = type;
-	e->base.send_level_to_reverb = 40.0f / 127.0f;
+	e->base.send_level_to_reverb = 40.0 / 127.0;
 	e->base.send_level_to_chorus = 0;
 	e->base.send_level_to_delay = 0;
 	e->base.process = phaw_process;
